% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tinyobjloader.R
\name{tinyobjloader-package}
\alias{tinyobjloader}
\alias{tinyobjloader-package}
\title{Wrapper for tiny_obj_loader single file C++ library}
\description{
see \url{https://github.com/syoyo/tinyobjloader}
}

